size_t to unsigned int

+++ includes/libmy.h
@@ -4,7 +4,7 @@
 #include <unistd.h>
 
 void my_putstr(const char *str);
-void *my_memcpy(void *dest, const void *src, size_t n);
+void *my_memcpy(void *dest, const void *src, unsigned int n);
 
 int my_is_alpha(char c);
 int my_is_digit(char c);
@@ -13,17 +13,17 @@ int my_tolower(int c);
 int my_toupper(int c);
 
 int my_strcmp(const char *s1, const char *s2);
-int my_strncmp(const char *s1, const char *s2, size_t n);
+int my_strncmp(const char *s1, const char *s2, unsigned int n);
 
-size_t my_count_words(const char *str, char sep);
-size_t my_strlen(const char *str);
-size_t my_strlcat(char *dest, const char *src, size_t size);
-size_t my_strlcpy(char *dest, const char *src, size_t size);
+unsigned int my_count_words(const char *str, char sep);
+unsigned int my_strlen(const char *str);
+unsigned int my_strlcat(char *dest, const char *src, unsigned int size);
+unsigned int my_strlcpy(char *dest, const char *src, unsigned int size);
 
 char *my_strcpy(char *dest, const char *src);
-char *my_strncpy(char *dest, const char *src, size_t n);
+char *my_strncpy(char *dest, const char *src, unsigned int n);
 char *my_strcat(char *dest, const char *src);
-char *my_strncat(char *dest, const char *src, size_t n);
+char *my_strncat(char *dest, const char *src, unsigned int n);
 
 char **my_split(const char *str, char sep);
