+++ my_strnlen.c
@@ -6,16 +6,16 @@
  * In doing this, strnlen() looks only at the first maxlen bytes at s and never beyond s+maxlen.
  *
  * @param str string
- * @param maxlen size_t
+ * @param maxlen unsigned int
  * 
  * @return The strnlen() function returns strlen(s), if that is less than maxlen,
  * or maxlen if there is no null byte ('\0') among the first maxlen bytes pointed to by s.
  *
  * @since 1.0
  */
-size_t my_strnlen(const char *str, size_t maxlen)
+unsigned int my_strnlen(const char *str, unsigned int maxlen)
 {
-    size_t i = 0;
+    unsigned int i = 0;
 
     while (str[i] && i < maxlen)
         i++;
